+++ common/engine/khg_utl/include/khg_utl/matrix.h
@@ -7,10 +7,10 @@
 typedef struct Matrix {
   unsigned int rows;
   unsigned int cols;
-  double *data;
+  float *data;
 } utl_matrix;
 
-typedef double (*matrix_func)(double); 
+typedef float (*matrix_func)(float); 
 
 utl_matrix *utl_matrix_create(unsigned int rows, unsigned int cols);
 utl_matrix *utl_matrix_add(const utl_matrix *matrix1, const utl_matrix *matrix2);
@@ -30,7 +30,7 @@ utl_matrix *utl_matrix_copy(const utl_matrix *matrix);
 utl_matrix *utl_matrix_kronecker_product(const utl_matrix *A, const utl_matrix *B);
 utl_matrix *utl_matrix_hankel(const utl_matrix *first_row, const utl_matrix *last_col);
 utl_matrix *utl_matrix_toeplitz(const utl_matrix *first_row, const utl_matrix *first_col);
-utl_matrix *utl_matrix_from_array(const double *data, unsigned int rows, unsigned int cols);
+utl_matrix *utl_matrix_from_array(const float *data, unsigned int rows, unsigned int cols);
 utl_matrix *utl_matrix_circulant(const utl_matrix *first_row);
 utl_matrix *utl_matrix_hilbert(unsigned int n);
 utl_matrix *utl_matrix_helmert(unsigned int n, bool full);
@@ -55,11 +55,11 @@ void utl_matrix_deallocate(utl_matrix *matrix);
 void utl_matrix_print(utl_matrix *matrix);
 void utl_matrix_swap_rows(utl_matrix *mat, unsigned int row1, unsigned int row2);
 void utl_matrix_swap_cols(utl_matrix *mat, unsigned int col1, unsigned int col2);
-void utl_matrix_row_divide(utl_matrix *matrix, unsigned int row, double scalar);
-void utl_matrix_row_subtract(utl_matrix *matrix, unsigned int target_row, unsigned int subtract_row, double scalar);
+void utl_matrix_row_divide(utl_matrix *matrix, unsigned int row, float scalar);
+void utl_matrix_row_subtract(utl_matrix *matrix, unsigned int target_row, unsigned int subtract_row, float scalar);
 
-bool utl_matrix_scalar_multiply(utl_matrix *matrix, double scalar);
-bool utl_matrix_set(utl_matrix *matrix, unsigned int rows, unsigned int cols, double value);
+bool utl_matrix_scalar_multiply(utl_matrix *matrix, float scalar);
+bool utl_matrix_set(utl_matrix *matrix, unsigned int rows, unsigned int cols, float value);
 bool utl_matrix_is_square(const utl_matrix *matrix);
 bool utl_matrix_is_equal(const utl_matrix *matrix1, const utl_matrix* matrix2);
 bool utl_matrix_is_identity(const utl_matrix *matrix);
@@ -77,23 +77,23 @@ bool utl_matrix_is_toeplitz(const utl_matrix *matrix);
 bool utl_matrix_lu_decomposition(const utl_matrix *matrix, utl_matrix **l, utl_matrix **u);
 bool utl_matrix_qr_decomposition(const utl_matrix *a, utl_matrix **q, utl_matrix **r);
 bool utl_matrix_is_positive_definite(const utl_matrix *matrix);
-bool utl_matrix_fill(utl_matrix *matrix, double value);
+bool utl_matrix_fill(utl_matrix *matrix, float value);
 bool utl_matrix_apply_to_row(utl_matrix *matrix, unsigned int row, matrix_func func);
 bool utl_matrix_apply_to_col(utl_matrix *matrix, unsigned int col, matrix_func func);
-bool utl_matrix_row_addition(utl_matrix *matrix, unsigned int target_row, unsigned int source_row, double scale);
-bool utl_matrix_col_addition(utl_matrix *matrix, unsigned int target_col, unsigned int source_col, double scale);
+bool utl_matrix_row_addition(utl_matrix *matrix, unsigned int target_row, unsigned int source_row, float scale);
+bool utl_matrix_col_addition(utl_matrix *matrix, unsigned int target_col, unsigned int source_col, float scale);
 bool utl_matrix_is_sparse(const utl_matrix *matrix);
 
-double utl_matrix_get(const utl_matrix *matrix, unsigned int row, unsigned int col);
-double utl_matrix_determinant(const utl_matrix *matrix);
-double utl_matrix_trace(const utl_matrix *matrix);
-double utl_matrix_frobenius_norm(const utl_matrix *matrix);
-double utl_matrix_l1_norm(const utl_matrix *matrix);
-double utl_matrix_infinity_norm(const utl_matrix *matrix);
-double utl_matrix_min_element(const utl_matrix *matrix);
-double utl_matrix_max_element(const utl_matrix *matrix);
+float utl_matrix_get(const utl_matrix *matrix, unsigned int row, unsigned int col);
+float utl_matrix_determinant(const utl_matrix *matrix);
+float utl_matrix_trace(const utl_matrix *matrix);
+float utl_matrix_frobenius_norm(const utl_matrix *matrix);
+float utl_matrix_l1_norm(const utl_matrix *matrix);
+float utl_matrix_infinity_norm(const utl_matrix *matrix);
+float utl_matrix_min_element(const utl_matrix *matrix);
+float utl_matrix_max_element(const utl_matrix *matrix);
 
-double *utl_matrix_to_array(const utl_matrix *matrix);
+float *utl_matrix_to_array(const utl_matrix *matrix);
 int utl_matrix_rank(const utl_matrix *matrix);
 unsigned int utl_matrix_size(const utl_matrix *matrix);
